requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Security
Werkzeug==3.0.1

# Environment Variables
python-dotenv==1.0.0

# Email Handling
email-validator==2.1.0

# HTTP Requests
requests==2.31.0

# AWS Bedrock (uncomment when migrating to Bedrock)
# boto3==1.34.0
# botocore==1.34.0

# Database
# SQLite is included in Python, no need to install

# Optional: For better development experience
# watchdog==3.0.0  # Auto-reload on file changes